% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{HDBSCAN}
\alias{HDBSCAN}
\title{Hierarchical Density-Based Spatial Clustering of Applications with Noise}
\usage{
HDBSCAN(object, ...)
}
\arguments{
\item{object}{input type}

\item{...}{passed on to specific HDBSCAN functions}
}
\value{
A list object (length=5) with the cluster labels for each point (labels [num]), the strength of a sample's membership to its assigned cluster (probabilities [num]), the stability of the cluster (cluster_persistance [num]),  list of exemplar points for clusters, and outlier scores for clusterd points (outlier_scores [num]))
}
\description{
Hierarchical Density-Based Spatial Clustering of Applications with Noise
}
\examples{
\dontrun{
# load Seurat object
data('pbmc_small', package = 'SeuratObject')

# tsne embedding matrix
emb <- Seurat::Embeddings(object = pbmc_small, 'tsne')

# HDBSCAN.matrix
res.matrix <- dbsinglecell::HDBSCAN(emb, min_cluster_size = 5)
res.matrix

# HDBSCAN.Seurat
res.seurat <- dbsinglecell::HDBSCAN(pbmc_small,reduction = 'tsne', min_cluster_size = 5)
slot(res.seurat, 'misc')$hdbscan

}

}
